refactoring reducers to pure functions
